% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-helpers.R, R/utils-helpers.R
\name{.gg_line_arg}
\alias{.gg_line_arg}
\title{Determine ggplot2 Line Width Argument Name by Version}
\usage{
.gg_line_arg()

.gg_line_arg()
}
\value{
Character string of argument name.

"linewidth" if ggplot2 >= 3.4.0, else "size".
}
\description{
Returns the correct argument name for line width in ggplot2,
depending on package version ("linewidth" for ≥ 3.4.0, else "size").
}
\keyword{internal}
