% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-helpers.R
\name{or_or}
\alias{or_or}
\alias{\%||\%}
\title{Internal Coalesce Operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Left-hand value.}

\item{y}{Right-hand value.}
}
\value{
If \code{x} is not NULL, returns \code{x}, else \code{y}.
}
\description{
Returns the left-hand side if it is not \code{NULL}, otherwise the right-hand side.
}
\keyword{internal}
