% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-bf.R
\name{plot_bf_heatmap}
\alias{plot_bf_heatmap}
\title{Plot Bayes Factor Heatmap (mean log10 BF10)}
\usage{
plot_bf_heatmap(power_results, title = NULL, subtitle = NULL)
}
\arguments{
\item{power_results}{Simulation results from a \verb{brms_inla_power*} function
with \code{compute_bayes_factor = TRUE}.}

\item{title, subtitle}{Optional ggplot title/subtitle.}
}
\value{
A ggplot object.
}
\description{
Filled contour plot of mean log10 BF10 as a function of
sample size and true effect size.
}
